+++ kif_lib/store/reader/reader.py
@@ -10,10 +10,24 @@ import functools
 import logging
 import pathlib
 
+from typing_extensions import overload
+
 from ... import itertools
-from ...model import Filter
+from ...context.registry import EntityRegistry
+from ...model import Entity, Filter
 from ...model import Graph as KIF_Graph
-from ...model import KIF_Object, Statement, TGraph
+from ...model import (
+    Item,
+    KIF_Object,
+    Lexeme,
+    Property,
+    Statement,
+    TDatatype,
+    TGraph,
+    TItem,
+    TProperty,
+    TText,
+)
 from ...typing import (
     Any,
     AsyncIterator,
@@ -27,10 +41,12 @@ from ...typing import (
     override,
     TextIO,
     TypeAlias,
+    TypeVar,
     Union,
 )
 from ..abc import Store
 
+E = TypeVar('E', bound=Entity)
 T: TypeAlias = Any
 
 _logger: Final[logging.Logger] = logging.getLogger(__name__)
@@ -57,6 +73,13 @@ class Reader(
     @dataclasses.dataclass
     class Source(abc.ABC):
         """Abstract base class for input sources."""
+
+    @dataclasses.dataclass
+    class Preamble(Source):
+        """Preamble source (sentinel)."""
+
+    class Postamble(Source):
+        """Postamble source (sentinel)."""
 
     @dataclasses.dataclass
     class Location(Source):
@@ -93,6 +116,8 @@ class Reader(
         '_args',
         '_kwargs',
         '_parse_fn',
+        '_registered',
+        '_registry',
     )
 
     #: Input sources.
@@ -103,6 +128,12 @@ class Reader(
 
     #: Parsing function.
     _parse_fn: ParseFn
+
+    #: Registered entities.
+    _registered: set[Entity]
+
+    #: Entity registry.
+    _registry: EntityRegistry
 
     def __init__(
             self,
@@ -127,6 +158,7 @@ class Reader(
                 raise KIF_Object._arg_error(
                     str(err), type(self), name,
                     exception=self.Error) from err
+        self._push_preamble()
         if location is not None:
             push('location', self._push_location, location)
         if file is not None:
@@ -143,10 +175,13 @@ class Reader(
                 *cast(Iterable[Statement], stmts)))
         for src in other:
             push('args', self._push_arg, src)
+        self._push_postamble()
         if parse is None:
             self._parse_fn = self._parse
         else:
             self._parse_fn = parse
+        self._registered = set()
+        self._registry = EntityRegistry()
 
     def _push_arg(self, arg: Any) -> None:
         if isinstance(arg, (pathlib.PurePath, str)):
@@ -165,6 +200,12 @@ class Reader(
             f'{type(self).__qualname__} does not support '
             f'{type(arg).__qualname__}')
 
+    def _push_preamble(self) -> None:
+        self._args.append(self.Preamble())
+
+    def _push_postamble(self) -> None:
+        self._args.append(self.Postamble())
+
     def _push_location(self, location: pathlib.PurePath | str) -> None:
         self._args.append(self.Location(location))
 
@@ -177,12 +218,128 @@ class Reader(
     def _push_graph(self, graph: KIF_Graph) -> None:
         self._args.append(self.Graph(graph))
 
+    @overload
+    def _register(self, entity: Item, **kwargs: Any) -> Item:
+        ...
+
+    @overload
+    def _register(self, entity: Property, **kwargs: Any) -> Property:
+        ...
+
+    @overload
+    def _register(self, entity: Lexeme, **kwargs: Any) -> Lexeme:
+        ...
+
+    def _register(
+            self,
+            entity: E,
+            label: TText | None = None,
+            labels: Iterable[TText] | None = None,
+            alias: TText | None = None,
+            aliases: Iterable[TText] | None = None,
+            description: TText | None = None,
+            descriptions: Iterable[TText] | None = None,
+            range: TDatatype | None = None,
+            inverse: TProperty | None = None,
+            lemma: TText | None = None,
+            category: TItem | None = None,
+            language: TItem | None = None,
+            **kwargs: Any
+    ) -> E:
+        self._registered.add(entity)
+        return self._registry.register(  # type: ignore
+            entity,                      # type: ignore
+            label=label,
+            labels=labels,
+            alias=alias,
+            aliases=aliases,
+            description=description,
+            descriptions=descriptions,
+            range=range,
+            inverse=inverse,
+            lemma=lemma,
+            category=category,
+            language=language,
+            function=self._register,
+            **kwargs)
+
+    @overload
+    def _register_descriptor(
+            self,
+            entity: Item,
+            descriptor: Item.Descriptor | None = None
+    ) -> Item:
+        ...
+
+    @overload
+    def _register_descriptor(
+            self,
+            entity: Property,
+            descriptor: Property.Descriptor | None = None
+    ) -> Property:
+        ...
+
+    @overload
+    def _register_descriptor(
+            self,
+            entity: Lexeme,
+            descriptor: Lexeme.Descriptor | None = None
+    ) -> Lexeme:
+        ...
+
+    def _register_descriptor(
+            self,
+            entity: E,
+            descriptor: Item.Descriptor
+            | Property.Descriptor | Lexeme.Descriptor | None = None
+    ) -> E:
+        assert isinstance(entity, (Item, Property, Lexeme))
+        if descriptor is None:
+            descriptor = entity.describe()
+        if descriptor is not None:
+            if isinstance(entity, (Item, Property)):
+                t = cast(
+                    Union[Item.Descriptor, Property.Descriptor], descriptor)
+                if 'labels' in t:
+                    self._register(entity, labels=t['labels'].values())
+                if 'aliases' in t:
+                    self._register(
+                        entity, aliases=itertools.chain(
+                            *t['aliases'].values()))
+                if 'description' in t:
+                    self._register(
+                        entity, description=t['descriptions'].values())
+                if isinstance(entity, Property):
+                    t = cast(Property.Descriptor, descriptor)
+                    if 'range' in t:
+                        self._register(entity, range=t['range'])
+                    if 'inverse' in t:
+                        self._register(entity, inverse=t['inverse'])
+            elif isinstance(entity, Lexeme):
+                l = cast(Lexeme.Descriptor, descriptor)
+                if 'lemma' in l:
+                    self._register(entity, lemma=l['lemma'])
+                if 'category' in l:
+                    self._register(entity, category=l['category'])
+                if 'language' in l:
+                    self._register(entity, language=l['language'])
+        return entity           # type: ignore
+
     def _load(self, file: TextIO) -> Iterable[T]:
         return file
+
+    def _preamble(self) -> Iterable[Statement]:
+        return iter(())
+
+    def _describe(self) -> Iterable[Entity]:
+        return iter(())
 
     def _parse(self, input: T) -> Iterable[Statement]:
         assert isinstance(input, str)
         yield Statement.from_json(input, **self._kwargs)
+
+    def _postamble(self) -> Iterable[Statement]:
+        return iter(())
 
     @override
     def _filter(
@@ -206,7 +363,19 @@ class Reader(
         parse = functools.partial(self._filter_parse, filter, options)
         wrap = functools.partial(self._filter_parse_wrap, filter, options)
 
-        if isinstance(arg, self.Location):
+        if isinstance(arg, self.Preamble):
+            yield from wrap(iter(self._preamble()))
+            for entity in self._describe():
+                assert isinstance(entity, (Item, Property, Lexeme))
+                self._register_descriptor(entity)
+        elif isinstance(arg, self.Postamble):
+            for entity in self._registered:
+                f = functools.partial(Statement, entity)
+                it = entity.descriptor_to_snaks(  # type: ignore
+                    self._registry.describe(entity))  # type: ignore
+                yield from wrap(map(f, it))
+            yield from wrap(iter(self._postamble()))
+        elif isinstance(arg, self.Location):
             with open(arg.location, encoding='utf-8') as fp:
                 yield from wrap(parse(fp))
         elif isinstance(arg, self.File):
@@ -269,8 +438,12 @@ class Reader(
             return await asyncio.to_thread(
                 lambda: (it, itertools.take(options.page_size, it)))
 
-        it = self._afilter_helper(list(map(parse, self._args)), task)
-        return itertools.amix(it, distinct=distinct, limit=limit)
+        assert len(self._args) >= 3
+        pre, mid, pos = (self._args[0],), self._args[1:-1], (self._args[-1],)
+        f = (lambda it: self._afilter_helper(list(map(parse, it)), task))
+        return itertools.amix(
+            f(pre), f(mid), f(pos),
+            distinct=distinct, limit=limit, method='chain')
 
     async def _afilter_helper(
             self,